Finish and test link_strategy in new linker
